Move Entity Models to Data/Entities, Create BaseEntitiy
This will keep shared logic in BaseEntity, and keep Entity Models separate from View Models.
