+++ FocusFlow/Data/AppDbContext.cs
@@ -1,6 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 
-using Task = FocusFlow.Models.Task;
+using Task = FocusFlow.Data.Entities.Task;
 
 namespace FocusFlow.Data;
