+++ FocusFlow/Data/Entities/Task.cs
@@ -1,8 +1,7 @@
-namespace FocusFlow.Models;
+namespace FocusFlow.Data.Entities;
 
-public sealed class Task
+public sealed class Task : BaseEntity
 {
-    public int Id { get; init; }
     public string Title { get; init; } = string.Empty;
     public string Description { get; init; } = string.Empty;
     public bool IsCompleted { get; init; } = false;
